December 2025 Update
It's been a bit since i've updated this repo, i've added a ton of notes from my last semester but also from starting work.

I've also been trying to go through and add tags and fix the maps and related notes section of my notes and i'm only about 40% done so there may be a new years update with the rest of  it.